postgresql changes commit

+++ src/main/resources/db/migration/V1_0__create_employees_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE employees
 (
-id bigint PRIMARY KEY,
+id SERIAL PRIMARY KEY,
 first_name varchar(50),
 last_name varchar(50),
 email varchar(50)
